+++ src/pages/posts/components/NoPosts/index.ts
@@ -1,0 +1,3 @@
+import NoPosts from './NoPosts';
+
+export { NoPosts };
