+++ src/pages/posts/components/PostContents/PostContents.style.ts
@@ -1,0 +1,3 @@
+import styled from '@emotion/styled';
+
+export const PostsContainer = styled.div``;
