+++ src/pages/posts/components/PostContents/index.ts
@@ -1,0 +1,4 @@
+import PostContents from './PostContents';
+import PostPreviewList from './PostPreviewList';
+
+export { PostContents, PostPreviewList };
