--- src/pages/posts/components/PostsItems/index.ts
@@ -1,3 +1,0 @@
-import PostItems from './PostItems';
-
-export { PostItems };
